+++ src/import-data.js
@@ -2,30 +2,8 @@ const asyncModule = require("async");
 const clear = require("clear");
 const archiveClear = require("./database-prep/archive-clear");
 const folderIntl = require("./file-prep/folder-intl");
-const importPaths = require("./import-export/import-paths");
-const importInfo = require("./import-export/import-info");
 const readImportFiles = require("./import-export/read-import-files");
 const exitProgram = require("./common/exit-program");
-
-
-/*
-	Command: import-data
-	
-	Description:
-		Imports exported .csv files into database.
-		This will completely overwrite any existing data.
-		
-	Steps:
-		* Validate target folder path input.
-		* Read 'ignore-errors' input.
-		* Remove existing data from archive database. (See 'clear-database.js')
-		* Initialize target folder.
-		* Check if import file paths exist.
-		* Read Advertiser data file and insert contents.
-		* Read ProductCategory data file and insert contents.
-		* Read MediaType data file and insert contents.
-		* Read CaseFile data file and insert contents.
-*/