+++ changelog.md
@@ -1,4 +1,14 @@
 # Changelog
 
-**./src/export-data.js**
-* Deleted file.
+**./src/import-export/**
+* Deleted files:
+	* import-info.js
+	* import-paths.js
+
+---
+
+**./src/import-data.js**
+* Removed requirements:
+	* ./import-export/import-paths
+	* ./import-export/import-info
+* Removed header comment.